Gravity now has frameTime  Large amount of notes on this github push read them
I tested all the maps they load and jump the same on my computer
-level1.txt the player falls through that is because I think is when it loads the map it is doing same loop were all the player functions are it loads it in too fast or slow but on the others the player is on the block or abit above dropping onto the block
suggestion the Menu needs to have its own gameloop to stop this so then we know it is loading the map on time.

So load menu game loop first once the map is selected it exits the menu game loop Loads the Map then enters the new game loop for the game.